Split unit tests into multiple files
